I updated the program to open the images using the PIL library, and then write the file names to a dictionary with an integer as the key so that I can apply my math.
Currently, it works with all of the images, except for the last image it
opens. I get a permission error when trying to rename file 99, stating
that the file is open in another spot. I tried adding time to let PIL
finish what it needs to do to close the file before rewriting, but it was
unsuccessful. Any advice is greatly appreciated!